refactor(channels): move discord_post_gate -> channels/discord/post_gate
First slice of the channels/discord/ lane restructure (phase-1a). Pure move +
dotted-import rewrite across all 5 production senders; no content change.
