+++ skills/bot2bot-post/post.py
@@ -265,7 +265,7 @@ _repo = next(p for p in pathlib.Path(__file__).resolve().parents
              if (p / "src" / "body_file.py").is_file())
 sys.path.insert(0, str(_repo / "src"))
 from body_file import MAX_BODY_BYTES, read_body_file as _read_body_file  # noqa: E402
-from discord_post_gate import make_client  # noqa: E402  — the one Discord POST chokepoint
+from channels.discord.post_gate import make_client  # noqa: E402  — the one Discord POST chokepoint
 from outbox import DeliveryOutcome  # noqa: E402
