+++ src/channels/__init__.py
@@ -1,0 +1,1 @@
+"""AG2 Space / Sutando channel adapters."""
