+++ src/channels/discord/__init__.py
@@ -1,0 +1,1 @@
+"""Discord channel: API mechanics + injected post-gate (bridge, client, http, reader, post_gate)."""
