+++ src/dm-result.py
@@ -57,7 +57,7 @@ from send_allowlist import (  # noqa: E402
     SEND_ALLOWED_ROOTS as _SEND_ALLOWED_ROOTS,
 )
 from message_chunking import chunk_message, _is_fence_open_line  # noqa: E402  (Result Router S3 — shared fence-aware chunker; was a 4th private copy)
-from discord_post_gate import make_client  # noqa: E402  — shared transport + injected post-gate
+from channels.discord.post_gate import make_client  # noqa: E402  — shared transport + injected post-gate
 from outbox import DeliveryOutcome  # noqa: E402
